Incorporate cy_build.py directly within setup.py
It's been "out of sight, out of mind" for too long.
(is_pip_install() and _get_egg_name() have been unused since 2017.)
