+++ setup.py
@@ -35,7 +35,14 @@ from distutils.command.build import build
 from setuptools.command.sdist import sdist
 from distutils.errors import LinkError
 
-from cy_build import CyExtension as Extension, cy_build_ext as build_ext
+try:
+    from Cython.Distutils import build_ext
+except ImportError:
+    from setuptools.command.build_ext import build_ext
+
+from distutils.extension import Extension
+from distutils.sysconfig import get_config_var, get_config_vars
+
 try:
     import cython  # noqa
     HAVE_CYTHON = True
@@ -233,6 +240,70 @@ class cythonize_sdist(sdist):
         from Cython.Build import cythonize
         cythonize(self.distribution.ext_modules)
         sdist.run(self)
+
+
+# Override Cythonised build_ext command to customise macOS shared libraries.
+
+if sys.platform == 'darwin':
+    config_vars = get_config_vars()
+    config_vars['LDSHARED'] = config_vars['LDSHARED'].replace('-bundle', '')
+    config_vars['SHLIB_EXT'] = '.so'
+
+
+class CyExtension(Extension):
+    def __init__(self, *args, **kwargs):
+        self._init_func = kwargs.pop("init_func", None)
+        self._prebuild_func = kwargs.pop("prebuild_func", None)
+        Extension.__init__(self, *args, **kwargs)
+
+    def extend_includes(self, includes):
+        self.include_dirs.extend(includes)
+
+    def extend_macros(self, macros):
+        self.define_macros.extend(macros)
+
+    def extend_extra_objects(self, objs):
+        self.extra_objects.extend(objs)
+
+
+class cy_build_ext(build_ext):
+
+    def build_extension(self, ext):
+
+        if isinstance(ext, CyExtension) and ext._init_func:
+            ext._init_func(ext)
+
+        if not self.inplace:
+            ext.library_dirs.append(os.path.join(self.build_lib, "pysam"))
+
+        if sys.platform == 'darwin':
+            # The idea is to give shared libraries an install name of the form
+            # `@rpath/<library-name.so>`, and to set the rpath equal to
+            # @loader_path. This will allow Python packages to find the library
+            # in the expected place, while still giving enough flexibility to
+            # external applications to link against the library.
+            relative_module_path = ext.name.replace(".", os.sep) + (get_config_var('EXT_SUFFIX') or get_config_var('SO'))
+            library_path = os.path.join(
+                "@rpath", os.path.basename(relative_module_path)
+            )
+
+            if not ext.extra_link_args:
+                ext.extra_link_args = []
+            ext.extra_link_args += ['-dynamiclib',
+                                    '-rpath', '@loader_path',
+                                    '-Wl,-headerpad_max_install_names',
+                                    '-Wl,-install_name,%s' % library_path,
+                                    '-Wl,-x']
+        else:
+            if not ext.extra_link_args:
+                ext.extra_link_args = []
+
+            ext.extra_link_args += ['-Wl,-rpath,$ORIGIN']
+
+        if isinstance(ext, CyExtension) and ext._prebuild_func:
+            ext._prebuild_func(ext, self.force)
+
+        build_ext.build_extension(self, ext)
 
 
 # Override build command to add extra build steps.
@@ -621,8 +692,8 @@ metadata = {
     'url': "https://github.com/pysam-developers/pysam",
     'packages': package_list,
     'requires': ['cython (>=0.29.12)'],
-    'ext_modules': [Extension(**opts) for opts in modules],
-    'cmdclass': {'build': extra_build, 'build_ext': build_ext, 'clean_ext': clean_ext, 'sdist': cythonize_sdist},
+    'ext_modules': [CyExtension(**opts) for opts in modules],
+    'cmdclass': {'build': extra_build, 'build_ext': cy_build_ext, 'clean_ext': clean_ext, 'sdist': cythonize_sdist},
     'package_dir': package_dirs,
     'package_data': {'': ['*.pxd', '*.h', 'py.typed', '*.pyi'], },
     # do not pack in order to permit linking to csamtools.so
